# agentbox-client/openapi/agentbox-manager.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SandboxDesiredState": {
        "enum": [
          "present",
          "released",
          "deleted"
        ],
        "title": "SandboxDesiredState",
        "type": "string"
      },
      "SandboxHandleResponse": {
        "additionalProperties": false,
        "properties": {
          "allocation_epoch": {
            "minimum": 0.0,
            "title": "Allocation Epoch",
            "type": "integer"
          },
          "allocation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allocation Id"
          },
          "allocation_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AllocationState"
              },
              {
                "type": "null"
              }
            ]
          },
          "desired_state": {
            "$ref": "#/components/schemas/SandboxDesiredState"
          },
          "logical_id": {
            "format": "uuid",
            "title": "Logical Id",
            "type": "string"
          },
          "operation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Id"
          },
          "profile": {
            "$ref": "#/components/schemas/ProfileRefModel"
          },
          "ready": {
            "title": "Ready",
            "type": "boolean"
          },
          "retry_after_ms": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After Ms"
          },
          "workload_kind": {
            "$ref": "#/components/schemas/WorkloadKind"
          }
        },
        "required": [
          "workload_kind",
          "logical_id",
          "desired_state",
          "profile",
          "allocation_state",
          "allocation_id",
          "allocation_epoch",
          "ready",
          "operation_id"
        ],
        "title": "SandboxHandleResponse",
        "type": "object"
      },
...
      "WorkloadKind": {
        "enum": [
          "workspace",
          "function"
        ],
        "title": "WorkloadKind",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "AgentBox",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/sandboxes/{workload_kind}/{logical_id}": {
      "delete": {
        "operationId": "delete_sandboxes_workload_kind_logical_id",
        "parameters": [
          {
            "in": "path",
            "name": "workload_kind",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/WorkloadKind"
            }
          },
          {
            "in": "path",
            "name": "logical_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Logical Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "deadline_at",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Deadline At",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Destroy Sandbox"
      },
      "get": {
        "operationId": "get_sandboxes_workload_kind_logical_id",
        "parameters": [
          {
            "in": "path",
            "name": "workload_kind",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/WorkloadKind"
            }
          },
          {
            "in": "path",
            "name": "logical_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Logical Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxHandleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Inspect Sandbox"
      },
      "put": {
        "operationId": "put_sandboxes_workload_kind_logical_id",
        "parameters": [
          {
            "in": "path",
            "name": "workload_kind",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/WorkloadKind"
            }
          },
          {
            "in": "path",
            "name": "logical_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Logical Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnsureSandboxRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxHandleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxHandleResponse"
                }
              }
            },
            "description": "Accepted"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ensure Sandbox"
      }
    },
    "/sandboxes/{workload_kind}/{logical_id}/directories": {
      "put": {
        "operationId": "put_sandboxes_workload_kind_logical_id_directories",
        "parameters": [
          {
            "in": "path",
            "name": "workload_kind",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/WorkloadKind"
            }
          },
          {
            "in": "path",
            "name": "logical_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Logical Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "maxLength": 4096,
              "minLength": 1,
              "pattern": "^/",
              "title": "Path",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "deadline_at",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Deadline At",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Directory"
      }
    },
    "/sandboxes/{workload_kind}/{logical_id}/files": {
      "delete": {
        "operationId": "delete_sandboxes_workload_kind_logical_id_files",
        "parameters": [
          {
            "in": "path",
            "name": "workload_kind",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/WorkloadKind"
            }
          },
          {
            "in": "path",
            "name": "logical_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Logical Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "maxLength": 4096,
              "minLength": 1,
              "pattern": "^/",
              "title": "Path",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "recursive",
            "required": false,
            "schema": {
              "default": false,
              "title": "Recursive",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "deadline_at",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Deadline At",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete File"
      },
      "get": {
        "operationId": "get_sandboxes_workload_kind_logical_id_files",
        "parameters": [
          {
            "in": "path",
            "name": "workload_kind",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/WorkloadKind"
            }
          },
          {
            "in": "path",
            "name": "logical_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Logical Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "maxLength": 4096,
              "minLength": 1,
              "pattern": "^/",
              "title": "Path",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "deadline_at",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Deadline At",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Files"
      }
    },
...
    "/sandboxes/{workload_kind}/{logical_id}/files:move": {
      "post": {
        "operationId": "post_sandboxes_workload_kind_logical_id_files_move",
        "parameters": [
          {
            "in": "path",
            "name": "workload_kind",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/WorkloadKind"
            }
          },
          {
            "in": "path",
            "name": "logical_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Logical Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveFileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Move File"
      }
    },
    "/sandboxes/{workload_kind}/{logical_id}/files:stat": {
      "get": {
        "operationId": "get_sandboxes_workload_kind_logical_id_files_stat",
        "parameters": [
          {
            "in": "path",
            "name": "workload_kind",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/WorkloadKind"
            }
          },
          {
            "in": "path",
            "name": "logical_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Logical Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "maxLength": 4096,
              "minLength": 1,
              "pattern": "^/",
              "title": "Path",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "deadline_at",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Deadline At",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileStatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stat File"
      }
    },
...
  }
}